Execution picture 2

+++ foundation/Foundation2/Product.cs
@@ -5,9 +5,9 @@ public class Product
   private string _name;
   private string _productID;
   private double _price;
-  private double _quantity;
+  private int _quantity;
 
-  public Product(string name, string productID, double price, double quantity)
+  public Product(string name, string productID, double price, int quantity)
   {
     _name = name;
     _productID = productID;
